+++ src/lib/api/schema/atoms/index.ts
@@ -1,11 +1,11 @@
-export * from './Age.schema';
-export * from './ClinicalFocus.schema';
+export * from './Age_new.schema';
+export * from './_ClinicalFocus.schema';
 export * from './LocationType.schema';
-export * from './Ethnicity.schema';
+export * from './_Ethnicity.schema';
 export * from './Gender.schema';
 export * from './Language.schema';
 export * from './LocationFacility.schema';
 export * from './LocationState.schema';
 export * from './TherapeuticModality.schema';
-export * from './DeliveryMethod.schema';
-export * from './EventDeliveryMethod.schema';
+export * from './_DeliveryMethod.schema';
+export * from './_EventDeliveryMethod.schema';
